Moved MVT Config and actions to mvt package
